create docker configuration

+++ java-server/src/main/java/com/github/genraven/genesys/domain/Setting.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Data
 @Document(collection = "settings")
 public class Setting {
@@ -13,7 +15,7 @@ public class Setting {
     }
 
     @Id
-    private Long id;
+    private BigInteger id;
     private String name;
     private boolean magic;
     private boolean current;
